type

+++ admin/delete_all.php
@@ -5,8 +5,12 @@ if (!isset($_SESSION['loggedin']) || isset($_SESSION['loggedin']) == false)
     header("Location: login.php");
 }else
 {
-    mysqli_query($connection,"DELETE FROM `places`");
-    header("Location: login.php");
+    if ($_POST['DA']='ok')
+    {
+        mysqli_query($connection,"DELETE FROM `places`");
+        header("Location: login.php");
+    }
+
 }
